reorganize App component/files
we had an old App.js that wasn't being used so I got rid of it and moved
AppRoot.js to App.js
